requirements.txt
fire==0.4.0
kaldiio==2.17.2
numpy==1.22.4
PyYAML==6.0
scipy==1.10.0
tableprint==0.9.1
torchnet==0.0.4
tqdm==4.66.3
scikit-learn
matplotlib==3.5.1
flake8==3.8.2
flake8-bugbear
flake8-comprehensions
flake8-executable
flake8-pyi==20.5.0
mccabe
h5py
pycodestyle==2.6.0
pyflakes==2.2.0
lmdb==1.3.0
onnxruntime
soundfile
pypeln==0.4.9
wandb
silero-vad
pre-commit==3.5.0
s3prl
hdbscan==0.8.37
umap-learn==0.5.6
